Delete and added

+++ src/main/java/com/companyname/javasampletest/entity/StudentEntity.java
@@ -11,13 +11,14 @@ import javax.persistence.Table;
 public class StudentEntity {
 	@Id
     @GeneratedValue(strategy=GenerationType.AUTO)
-    private Integer id;
+    private Long id;
     private String name;
     private String email;
-	public Integer getId() {
+	
+	public Long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getName() {
@@ -32,6 +33,7 @@ public class StudentEntity {
 	public void setEmail(String email) {
 		this.email = email;
 	}
+	
 
 
 }
